refactor web source sets in compose-ui
