+++ compose/ui/ui/src/uikitMain/kotlin/androidx/compose/ui/viewinterop/InteropWrappingView.uikit.kt
@@ -28,7 +28,6 @@ import platform.CoreGraphics.CGRectZero
 import platform.UIKit.UIResponder
 import platform.UIKit.UIView
 import platform.UIKit.UIViewController
-import platform.UIKit.setAccessibilityElements
 
 /**
  * On iOS [InteropView] is a [UIResponder], which is a base class for [UIView] and [UIViewController]
