+++ compose/ui/ui/src/wasmJsMain/kotlin/androidx/compose/ui/platform/PlatformClipboard.wasm.kt
@@ -40,7 +40,7 @@ private val isFallbackWriteTextApiAvailable: Boolean by lazy {
     isSecureContext && isFallbackWriteTextApiAvailable()
 }
 
-class WasmPlatformClipboard : Clipboard {
+private class WasmPlatformClipboard : Clipboard {
 
     private val browserClipboard by lazy {
         getW3CClipboard()
