--- compose/ui/ui/src/wasmJsMain/kotlin/androidx/compose/ui/window/ComposeWindow.wasmJs.kt
@@ -1,37 +1,0 @@
-/*
- * Copyright 2025 The Android Open Source Project
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package androidx.compose.ui.window
-
-import kotlinx.browser.document
-import org.jetbrains.skiko.wasm.onWasmReady
-import org.w3c.dom.DocumentReadyState
-import org.w3c.dom.LOADING
-
-internal actual fun onSkikoReady(block: () -> Unit) {
-    onWasmReady { block() }
-}
-
-internal actual fun onDomReady(block: () -> Unit) {
-    // https://developer.mozilla.org/en-US/docs/Web/API/Document/DOMContentLoaded_event
-    if (document.readyState == DocumentReadyState.Companion.LOADING) {
-        document.addEventListener("DOMContentLoaded", {
-            block()
-        })
-    } else {
-        block()
-    }
-}
