+++ compose/ui/ui/src/webMain/kotlin/androidx/compose/ui/window/ComposeWindow.w3c.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2024 The Android Open Source Project
+ * Copyright 2025 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -689,7 +689,7 @@ fun CanvasBasedWindow(
     )
 }
 
-internal actual fun InternalComposeViewport(
+internal fun InternalComposeViewport(
     viewportContainerId: String?,
     configure: ComposeViewportConfiguration.() -> Unit,
     content: @Composable () -> Unit
